Fixed integer underflow with call stack checker causing recursion depth error false positives

+++ include/cognate.c
@@ -152,10 +152,10 @@ static void check_call_stack()
   if (unlikely(calls > 1024))
   {
     calls = 0;
-    static unsigned long old_stack_size;
+    static long old_stack_size;
     char b;
     // if (how much stack left < stack change between checks)
-    if (stack_max.rlim_cur - (unsigned long)(stack_start - &b) < stack_start - &b - old_stack_size)
+    if ((long)stack_max.rlim_cur - (long)(stack_start - &b) < (long)stack_start - (long)&b - old_stack_size)
     {
       throw_error("Call stack overflow! Too much recursion!");
     }
